Gaussian,Vasp,MolVisPython: moved common functionality to MolVisPython
