+++ molvis/gaussian/python/gaussianutil.py
@@ -2,7 +2,7 @@
  #
  # Inviwo - Interactive Visualization Workshop
  #
- # Copyright (c) 2020-2021 Inviwo Foundation
+ # Copyright (c) 2020-2022 Inviwo Foundation
  # All rights reserved.
  #
  # Redistribution and use in source and binary forms, with or without
@@ -28,10 +28,8 @@
  #################################################################################
 
 import inviwopy as ivw
-import ivwdataframe as df
 import ivwmolvis
 
-import itertools
 import numpy
 
 from pathlib import Path
@@ -64,7 +62,7 @@ def parseCubeFile(file, flipSign=False, centerData=True):
         extraLines = 1
     if len(splitted) > 4:
         nVal = int(splitted[4])
-    
+
     unitsInAngstrom = False
     bohrToAngstrom = 0.529177249
     splitted = lines[3].strip().split()
@@ -90,7 +88,7 @@ def parseCubeFile(file, flipSign=False, centerData=True):
         origin = bohrToAngstrom * origin
 
     offset = -0.5 * (basis[0] + basis[1] + basis[2]) if centerData else 0
-    
+
     pos = []
     atoms = []
     for i in range(numAtoms):
@@ -99,7 +97,7 @@ def parseCubeFile(file, flipSign=False, centerData=True):
         atomPos = numpy.array(list(map(float, splitted[2:5])))
         if not unitsInAngstrom:
             atomPos = bohrToAngstrom * atomPos
-        atomPos = atomPos - origin 
+        atomPos = atomPos - origin
         pos.append(atomPos)
         atoms.append(ivwmolvis.atomicelement.element(atomID))
 
@@ -110,7 +108,7 @@ def parseCubeFile(file, flipSign=False, centerData=True):
         splitted = lines[i].strip().split()
         for val in splitted:
             chg.append(sign * float(val))
-    
+
     chgdata = numpy.array(chg).astype(numpy.float32)
     chosenIdx = 0
     chgdata = chgdata[chosenIdx::nVal]
@@ -137,93 +135,3 @@ def parseCubeFile(file, flipSign=False, centerData=True):
     volume.offset = ivw.glm.vec3(offset)
 
     return (volume, pos, atoms)
-
-def createMeshForCube(pos, elements, basis, offset, pm, margin, radiusScaling = 1.0):
-    position = []
-    color = []
-    radius = []
-    picking = []
-    index = []
-
-    pm.resize(len(elements))
-
-    for i, p in enumerate(pos):
-        c = numpy.array(ivwmolvis.atomicelement.color(elements[i]))
-        r = ivwmolvis.atomicelement.vdwRadius(elements[i]) * radiusScaling
-        pi = pm.pickingId(i)
-
-        def addVertex(vertexpos):
-            position.append(vertexpos)
-            color.append(c)
-            radius.append(r)
-            picking.append(pi)
-            index.append(i)
-
-        if margin > 0.0:
-            for shift in itertools.product([-1, 0, 1], repeat=3):
-                if all([x+s > -margin and x+s < 1 + margin for s, x in zip(shift, p)]):
-                    addVertex([x+s for s, x in zip(shift, p)])
-        else:
-            addVertex(p)
-
-
-    mesh = ivw.data.Mesh()
-    #mesh.basis = ivw.glm.mat3(basis)
-    mesh.offset = ivw.glm.vec3(offset)
-
-    mesh.addBuffer(ivw.data.BufferType.PositionAttrib, ivw.data.Buffer(
-        numpy.array(position).astype(numpy.float32)))
-    mesh.addBuffer(ivw.data.BufferType.ColorAttrib, ivw.data.Buffer(
-        numpy.array(color).astype(numpy.float32)))
-    mesh.addBuffer(ivw.data.BufferType.RadiiAttrib, ivw.data.Buffer(
-        numpy.array(radius).astype(numpy.float32)))
-    mesh.addBuffer(ivw.data.BufferType.PickingAttrib, ivw.data.Buffer(
-        numpy.array(picking).astype(numpy.uint32)))
-    mesh.addBuffer(ivw.data.BufferType.IndexAttrib, ivw.data.Buffer(
-        numpy.array(index).astype(numpy.uint32)))
-    return mesh
-
-def createMolecularStructure(pos, elements, margin, offset):
-    positions = []
-    indices = []
-    atomicnumbers = []
-    for i, p in enumerate(pos):
-        element = elements[i]
-
-        def addAtom(atompos):
-            positions.append(ivw.glm.dvec3(atompos) + offset)
-            indices.append(i)
-            atomicnumbers.append(element)
-
-        if margin > 0.0:
-            for shift in itertools.product([-1, 0, 1], repeat=3):
-                if all([x+s > -margin and x+s < 1 + margin for s, x in zip(shift, p)]):
-                    addAtom([x+s for s, x in zip(shift, p)])
-        else:
-            addAtom(p)
-    atoms = ivwmolvis.Atoms()
-    atoms.positions = positions
-    atoms.serialnumbers = indices
-    atoms.atomicnumbers = atomicnumbers
-    bonds = ivwmolvis.util.computeCovalentBonds(atoms)
-    return ivwmolvis.MolecularStructure(ivwmolvis.MolecularData(source="chgcar_file", atoms=atoms, 
-                                                                residues=[], chains=[], 
-                                                                bonds=bonds))
-def createDataFrameForCube(pos, elements):
-    dataframe = df.DataFrame()
-    ct = dataframe.addCategoricalColumn("type")
-    cx = dataframe.addFloatColumn("x", 0, ivw.data.Unit("Angstrom"))
-    cy = dataframe.addFloatColumn("y", 0, ivw.data.Unit("Angstrom"))
-    cz = dataframe.addFloatColumn("z", 0, ivw.data.Unit("Angstrom"))
-    r = dataframe.addFloatColumn("r", 0, ivw.data.Unit("Angstrom"))
-
-    for elem, p in zip(elements, pos):        
-        ct.add(ivwmolvis.atomicelement.symbol(elem))
-        cx.add(p[0])
-        cy.add(p[1])
-        cz.add(p[2])
-        r.add(ivwmolvis.atomicelement.vdwRadius(elem))
-
-    dataframe.updateIndex()
-
-    return dataframe
