+++ molvis/gaussian/python/processors/cubesource.py
@@ -4,7 +4,7 @@
  #
  # Inviwo - Interactive Visualization Workshop
  #
- # Copyright (c) 2020-2021 Inviwo Foundation
+ # Copyright (c) 2020-2022 Inviwo Foundation
  # All rights reserved.
  #
  # Redistribution and use in source and binary forms, with or without
@@ -32,6 +32,7 @@
 import inviwopy as ivw
 import ivwdataframe as df
 import ivwmolvis
+import molviscommon
 import gaussianutil
 
 import numpy as np
@@ -80,10 +81,6 @@ class CubeSource(ivw.Processor):
             "centerData", "Center Data", True)
         self.addProperty(self.centerData)
 
-        self.margin = ivw.properties.FloatProperty(
-            "margin", "Border Repetition Margin", 0.0, 0.0, 0.5, 0.01)
-        self.addProperty(self.margin)
-
         self.radiusScaling = ivw.properties.FloatProperty(
             "radiusScaling", "Radius Scaling", 0.25, 0.0, 2.0, 0.01)
         self.addProperty(self.radiusScaling)
@@ -118,15 +115,16 @@ class CubeSource(ivw.Processor):
         self.volume.dataMap.dataRange = self.customDataRange.value if self.useCustomRange.value else self.dataRange.value
         self.volume.dataMap.valueRange = self.volume.dataMap.dataRange
 
-        self.mesh = gaussianutil.createMeshForCube(self.atomPos, self.atomTypes,
-                                        self.volume.basis, self.volume.offset,
-                                        self.pm, self.margin.value,
-                                        self.radiusScaling.value)
+        self.mesh = molviscommon.createMesh(pos=self.atomPos, elements=self.atomTypes,
+                                            basis=self.volume.basis, offset=self.volume.offset,
+                                            pm=self.pm, radiusscaling=self.radiusScaling.value)
 
-        self.dataframe = gaussianutil.createDataFrameForCube(self.atomPos, self.atomTypes)
+        self.dataframe = molviscommon.createDataFrame(pos=self.atomPos, elements=self.atomTypes,
+                                                      modelmat=self.volume.modelMatrix)
 
         offset = ivw.glm.dvec3(self.volume.offset) if self.centerData.value else ivw.glm.dvec3(0)
-        self.molecule = gaussianutil.createMolecularStructure(self.atomPos, self.atomTypes, self.margin.value, offset)
+        self.molecule = molviscommon.createMolecularStructure(pos=self.atomPos, elements=self.atomTypes, 
+                                                              offset=offset)
         
         self.volumeOutport.setData(self.volume)
         self.meshOutport.setData(self.mesh)
